[WIP] Notification 2 examples
