--- hello-world-notification-microservice/src/main/java/c8y/example/notification/helloworld/HelloWorldApplication.java
@@ -1,11 +1,0 @@
-package c8y.example.notification.helloworld;
-
-import org.springframework.boot.SpringApplication;
-
-public class HelloWorldApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(HelloWorldConfiguration.class, args);
-    }
-
-}
