--- hello-world-notification-microservice/src/main/java/c8y/example/notification/helloworld/HelloWorldConfiguration.java
@@ -1,20 +1,0 @@
-package c8y.example.notification.helloworld;
-
-import com.cumulocity.microservice.autoconfigure.MicroserviceApplication;
-import com.cumulocity.microservice.subscription.model.MicroserviceMetadataRepresentation;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Primary;
-
-import static com.cumulocity.microservice.subscription.model.MicroserviceMetadataRepresentation.microserviceMetadataRepresentation;
-
-@MicroserviceApplication
-public class HelloWorldConfiguration {
-
-    @Bean
-    @Primary
-    public MicroserviceMetadataRepresentation helloWorldMicroserviceMetadata() {
-        return microserviceMetadataRepresentation()
-                .requiredRole("ROLE_NOTIFICATION_2_ADMIN")
-                .build();
-    }
-}
