--- hello-world-notification-microservice/src/main/java/c8y/example/notification/helloworld/websocket/WebSocketClient.java
@@ -1,6 +1,0 @@
-package c8y.example.notification.helloworld.websocket;
-
-public interface WebSocketClient extends AutoCloseable {
-
-    void connect() throws Exception;
-}
