--- hello-world-notification-microservice/src/main/resources/application.properties
@@ -1,10 +1,0 @@
-application.name=hello-notification
-server.port=8085
-
-example.source.id=29100
-example.websocket.url=ws://cumulocity.default.svc.cluster.local
-example.subscriber=testsubscriber
-
-# This parameter is provided by platform for microservice deployed in cumulocity.
-# You can set it up to your dedicated cumulocity address. You can get one by trying free trial on www.cumulocity.com
-#C8Y.baseURL=http://cumulocity.default.svc.cluster.local
