+++ notification2-examples/hello-world-microservice/src/main/configuration/cumulocity.json
