+++ notification2-examples/hello-world-microservice/src/main/java/c8y/example/notification/microservice/HelloWorldApplication.java
@@ -1,0 +1,11 @@
+package c8y.example.notification.microservice;
+
+import org.springframework.boot.SpringApplication;
+
+public class HelloWorldApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(HelloWorldConfiguration.class, args);
+    }
+
+}
