+++ notification2-examples/java-examples/pom.xml
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>c8y.example</groupId>
+        <artifactId>notification2-examples</artifactId>
+        <version>${revision}${changelist}</version>
+    </parent>
+
+    <artifactId>java-examples</artifactId>
+    <name>Cumulocity :: Examples :: Notification 2.0 Examples :: Java Examples</name>
+
+    <properties>
+        <commons-configuration2.version>2.9.0</commons-configuration2.version>
+    </properties>
+
+    <dependencies>
+        <dependency>
+            <groupId>com.nsn.cumulocity.clients-java</groupId>
+            <artifactId>java-client</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.commons</groupId>
+            <artifactId>commons-configuration2</artifactId>
+            <version>${commons-configuration2.version}</version>
+        </dependency>
+    </dependencies>
+
+</project>
