+++ notification2-examples/java-examples/src/main/java/c8y/example/notification/Example1.java
@@ -1,0 +1,11 @@
+package c8y.example.notification;
+
+import org.apache.commons.configuration2.ex.ConfigurationException;
+
+public class Example1 {
+    public static void main(String[] args) throws ConfigurationException {
+        final Properties properties = new Properties();
+        System.out.println(properties.getSourceId());
+    }
+
+}
