+++ notification2-examples/java-examples/src/main/resources/notifications-example.properties
@@ -1,0 +1,3 @@
+example.source.id=883202
+example.websocket.url=ws://cumulocity.default.svc.cluster.local
+example.subscriber=testsubscriber
