+++ notification2-examples/pom.xml
@@ -1,0 +1,44 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <groupId>c8y.example</groupId>
+        <artifactId>c8y-examples</artifactId>
+        <version>${revision}${changelist}</version>
+    </parent>
+
+    <artifactId>notification2-examples</artifactId>
+    <name>Cumulocity :: Examples :: Notification 2.0 Examples</name>
+    <packaging>pom</packaging>
+
+    <modules>
+        <module>hello-world-microservice</module>
+        <module>shared-components</module>
+        <module>java-examples</module>
+    </modules>
+
+    <properties>
+        <c8y.version>${revision}${changelist}</c8y.version>
+        <java.version>11</java.version>
+        <maven.compiler.target>${java.version}</maven.compiler.target>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+    </properties>
+
+    <dependencies>
+        <dependency>
+            <groupId>org.projectlombok</groupId>
+            <artifactId>lombok</artifactId>
+            <version>1.18.20</version>
+            <scope>provided</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.java-websocket</groupId>
+            <artifactId>Java-WebSocket</artifactId>
+            <version>1.5.1</version>
+        </dependency>
+    </dependencies>
+
+</project>
