+++ notification2-examples/shared-components/pom.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <groupId>c8y.example</groupId>
+        <artifactId>notification2-examples</artifactId>
+        <version>${revision}${changelist}</version>
+    </parent>
+
+    <artifactId>shared-components</artifactId>
+    <name>Cumulocity :: Examples :: Notification 2.0 Examples :: Shared Components</name>
+
+    <dependencies>
+        <dependency>
+            <groupId>com.nsn.cumulocity.clients-java</groupId>
+            <artifactId>java-client</artifactId>
+        </dependency>
+    </dependencies>
+
+</project>
