+++ notification2-examples/shared-components/src/main/java/c8y/example/common/platform/SubscriptionRepository.java
@@ -1,0 +1,26 @@
+package c8y.example.common.platform;
+
+import com.cumulocity.rest.representation.reliable.notification.NotificationSubscriptionRepresentation;
+import com.cumulocity.sdk.client.messaging.notifications.NotificationSubscriptionApi;
+import com.cumulocity.sdk.client.messaging.notifications.NotificationSubscriptionCollection;
+import com.cumulocity.sdk.client.messaging.notifications.NotificationSubscriptionFilter;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+public class SubscriptionRepository {
+
+    private final NotificationSubscriptionApi subscriptionApi;
+
+    public NotificationSubscriptionRepresentation create(NotificationSubscriptionRepresentation subscriptionRepresentation) {
+        return subscriptionApi.subscribe(subscriptionRepresentation);
+    }
+
+    public void delete(NotificationSubscriptionRepresentation subscriptionRepresentation) {
+        subscriptionApi.delete(subscriptionRepresentation);
+    }
+
+    public NotificationSubscriptionCollection getByFilter(NotificationSubscriptionFilter filter) {
+        return subscriptionApi.getSubscriptionsByFilter(filter);
+    }
+
+}
