+++ notification2-examples/shared-components/src/main/java/c8y/example/common/websocket/WebSocketClient.java
@@ -1,0 +1,6 @@
+package c8y.example.common.websocket;
+
+public interface WebSocketClient extends AutoCloseable {
+
+    void connect() throws Exception;
+}
