+++ pom.xml
@@ -53,7 +53,7 @@
   <modules>
     <module>tracker-agent</module>
     <module>hello-world-microservice</module>
-    <module>hello-world-notification-microservice</module>
+    <module>notification2-examples</module>
     <module>sample-lwm2m-custom-decoder</module>
     <module>sample-hex-decoder</module>
     <module>sample-json-decoder</module>
